requirements.txt
python-telegram-bot==21.6
anthropic==0.34.0
apscheduler==3.10.4
pytz==2024.1
